Fix get_balance
Rostrum 9.0.0 and Fulcrum 1.9.X introduced a modification to
blockchain.address.get_balance.

The change is related to the activation of CashTokens on May 2023

An optional parameter has been added to get_balance, namely token_filter,
so that in the calculated amount UTXOs associated to an address could be
counted in or not on the base of their association to a toke.

That lead to some kind of weird corner case where not all UTXOs got
considered while computing the amount of a given address.

The fix comes down to negotiate the proper protocol version between
the explorer and the electrum server. In case of Rostrum 1.4 is the max
proto version and the one that is needed to get a balance accounting
for all UTXOs. For Fulcrum 1.9.x 1.5 is the right proto version (prior
to that the get_balance is supposed to accept only 1 parameter).