Add a temporary rate limit to /address views
That view rendering could pretty resource consuming when
the address we are looking for has a very long transaction
history.

I'm going to temporarely add a rate limit for that view that
take into account the number of requests for a given time interval
by the same IP address.

Once pagination would be provided by Eletrum servers (eg Rostrum)
for the used methods (eg get_history)  we could use it to reduce
the amount of load anda data involved and we will be able to remove
the rate limit on the page.